iperf fix

+++ iperf/cilent.py
@@ -6,28 +6,42 @@ from common import *
 IPERF_PATH = 'iperf3/iperf3.exe'
 
 class ClientTcp(Producer):
-    def __init__(self, server_ip, server_port, duration, num_streams, reverse, output=sys.stderr):
+    def __init__(self, server_ip, server_port, duration, num_streams, reverse, bandwidth, output=sys.stderr, capture_output=False):
         self.server_ip = server_ip
         self.server_port = server_port
         self.duration = duration
         self.num_streams = num_streams
         self.reverse = reverse
+        self.bandwidth = bandwidth
         self.f = output
         self.process = None
+        self.capture_output = capture_output
         super().__init__()
+    
+    def repr(self):
+        return f"TCP {'⬇️' if self.reverse else '⬆️'}"
 
     def update(self):
         cmd = [
             IPERF_PATH, '-c', self.server_ip, '-p', str(self.server_port),
-            '-t', str(self.duration), '-P', str(self.num_streams), '-J'
+            '-t', str(self.duration), '-P', str(self.num_streams)
         ]
+        if len(self.bandwidth) > 0:
+            cmd.append('-b')
+            cmd.append(self.bandwidth)
         if self.reverse:
             cmd.append('-R')
-        self.f.write(f"[iperf]运行 {' '.join(cmd)}\n")
-        self.process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
-        stdout, stderr = self.process.communicate()
-        self.res = self.parse_output(stdout)
-        self.process.terminate()
+        if self.capture_output:
+            cmd.append('-J')
+            self.f.write(f"[iperf]运行 {' '.join(cmd)}\n")
+            self.process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
+            stdout, stderr = self.process.communicate()
+            self.res = (self.repr(), self.parse_output(stdout))
+            self.process.terminate()
+        else:
+            self.f.write(f"[iperf]运行 {' '.join(cmd)}\n")
+            self.process = subprocess.Popen(cmd, text=True)
+            self.res = (self.repr(), "不捕获输出")
 
     def parse_output(self, output):
         try:
@@ -41,7 +55,7 @@ class ClientTcp(Producer):
             return None
 
 class ClientUdp(Producer):
-    def __init__(self, server_ip, server_port, duration, num_streams, reverse, bandwidth, output=sys.stderr):
+    def __init__(self, server_ip, server_port, duration, num_streams, reverse, bandwidth, output=sys.stderr, capture_output=False):
         self.server_ip = server_ip
         self.server_port = server_port
         self.duration = duration
@@ -50,21 +64,32 @@ class ClientUdp(Producer):
         self.bandwidth = bandwidth
         self.f = output
         self.process = None
+        self.capture_output = capture_output
         super().__init__()
+    
+    
+    def repr(self):
+        return f"UDP {'⬇️' if self.reverse else '⬆️'}"
 
     def update(self):
         cmd = [
             IPERF_PATH, '-c', self.server_ip, '-p', str(self.server_port),
             '-t', str(self.duration), # '-P', str(self.num_streams),
-            '-u', '-b', str(self.bandwidth), '-J'
+            '-u', '-b', str(self.bandwidth)
         ]
         if self.reverse:
             cmd.append('-R')
-        self.f.write(f"[iperf]运行 {' '.join(cmd)}\n")
-        self.process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
-        stdout, stderr = self.process.communicate()
-        self.res = self.parse_output(stdout)
-        self.process.terminate()
+        if self.capture_output:
+            cmd.append('-J')
+            self.f.write(f"[iperf]运行 {' '.join(cmd)}\n")
+            self.process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
+            stdout, stderr = self.process.communicate()
+            self.res = (self.repr(), self.parse_output(stdout))
+            self.process.terminate()
+        else:
+            self.f.write(f"[iperf]运行 {' '.join(cmd)}\n")
+            self.process = subprocess.Popen(cmd, text=True)
+            self.res = (self.repr(), "不捕获输出")
     
     def parse_output(self, output):
         try:
